app/requirements.txt
Flask
pymysql
mysql-connector-python
